issue fixes

+++ src/app/components/academic-year/home/home.component.ts
@@ -163,14 +163,14 @@ export class HomeComponent implements OnInit {
 
   }
   editRowTable(row, index) {
-    
+
     document.getElementById(("row" + index).toString()).classList.remove('displayComp');
     document.getElementById(("row" + index).toString()).classList.add('editComp');
-    
+
   }
 
   cancelEditRow(index) {
-    
+
     document.getElementById(("row" + index).toString()).classList.add('displayComp');
     document.getElementById(("row" + index).toString()).classList.remove('editComp');
     this.getAllAcademicFromServer();
@@ -224,10 +224,10 @@ export class HomeComponent implements OnInit {
 
       this.academicyearservice.editAcademicYear(data, row2.inst_acad_year_id).subscribe(
         res => {
-          
+
           this.cancelEditRow(index);
           this.getAllAcademicFromServer();
-          
+
         },
         error => {
           let acad = {
@@ -242,28 +242,30 @@ export class HomeComponent implements OnInit {
     }
   }
 
-  deleteAcademicYear(row){
+  deleteAcademicYear(row) {
     let inst_id = row.inst_acad_year_id
-    this.academicyearservice.deleteAcademicYear(inst_id).subscribe(
-      (data:any)=>{
-        let msg = {
-          type:'success',
-          body:'Academic year deleted successfully'
+    if (confirm('Are you sure, you want to delete?')) {
+      this.academicyearservice.deleteAcademicYear(inst_id).subscribe(
+        (data: any) => {
+          let msg = {
+            type: 'success',
+            body: 'Academic year deleted successfully'
+          }
+          this.appC.popToast(msg);
+          this.getAllAcademicFromServer();
+        },
+        (error: any) => {
+          let msg = {
+            type: 'error',
+            body: error.error.message
+          }
+          this.appC.popToast(msg);
         }
-        this.appC.popToast(msg);
-        this.getAllAcademicFromServer();
-      },
-      (error:any)=>{
-        let msg = {
-          type:'error',
-          body:error.error.message
-        }
-        this.appC.popToast(msg);
-      }
-    )
+      )
+    }
   }
 
-  
+
   toggleCreateNewAcademicYear() {
     if (this.createNewAcademicYear == false) {
       this.createNewAcademicYear = true;
